+++ backend/HomeServer-Backend-win/Core/Endpoints/Endpoints_DELETE.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HomeServer_Backend.Core
+{
+    public partial class ServerCore
+    {
+
+    }
+}
